45: Aspect ratio updated at start (#46)
This change fixes the issues with the framebuffer not being updated at the start of the app. As it turned out, it was a macOS issue only, caused by the Retina rescaling. Therefore, this feature got disabled by `GLFW_COCOA_RETINA_FRAMEBUFFER` being set to false.  
Apart from that, the change provides minor fixes reported by clang-tidy:  
- updates printf format specifiers from %d to %zu for printing std::size_t variables  
- removes numerous unused header includes across the codebase.  

It also adds CMake configuration for debug builds and VS Code integration (added launch and tasks to build and debug using the VS Code interface).